Replace all options strings with constants from header
This makes the output filenames generated by unpackbootimg match the
input arguments to mkbootimg.  The presence/absence of an underscore
isn't a huge difference, but is a major pain when manually modifying
boot images.

This is a massive user interface improvement, at the cost of needing
older scripts/programs to be updated.