Rework write()/write_padding() calls
The calls to write_padding() were always paired with write() calls of the
exact same format.  Merge these pairs together into write_padded() and
thus shrink the code.

Redo the mechanism write_padded() uses to generate the pad.  Instead of
a write() of a statically allocated buffer, use a lseek() and
ftruncate().  This removes the need to always have a large static pad
hanging around and allows the code to support any page size which doesn't
overflow the integer types.