Merge checks, error message into load_file()
Every call to load_file() produces almost the same error in case of
failure.  Add this as a parameter to load_file() and a bunch of lines can
be removed.  Nearly every call to load_file() is optional if the file
isn't provided as an argument.  As such let load_file() simply ignore
empty filenames (though return zero).