Implement parameter files
There are enough paramters for a boot image file to be troublesome to
type them all on a command-line.  As such make it possible to dump them
all in a combined file, instead of needing separate arguments for each.

Spot and remove an extraneous newline from the header_version value.
In a text file this is harmless, in a paramter file this is far more
visible.